requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.1.0
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.11.0
Cython==0.29.24
dataclasses==0.8
decorator==4.4.2
distro==1.6.0
einops==0.3.2
fire==0.4.0
frozenlist==1.2.0
fsspec==2021.11.0
future==0.18.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.42.0
gspread==4.0.1
httplib2==0.20.2
idna==3.3
idna-ssl==1.1.0
imageio==2.11.1
importlib-metadata==4.8.2
joblib==1.1.0
kiwisolver==1.3.1
Markdown==3.3.6
matplotlib==3.3.4
multidict==5.2.0
networkx==2.5.1
nsml-watchdog==0.0.4
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.1
opencv-python==4.5.3.56
opencv-python-headless==4.5.4.60
packaging==21.3
Pillow==8.4.0
pkg_resources==0.0.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pyDeprecate==0.3.1
pyparsing==3.0.6
python-dateutil==2.8.2
pytorch-lightning==1.5.2
PyWavelets==1.1.1
PyYAML==6.0
qudida==0.0.4
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-build==0.12.0
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
threadpoolctl==3.0.0
tifffile==2020.9.3
torch==1.10.0
torchmetrics==0.6.0
torchvision==0.11.1
tqdm==4.62.3
typing_extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
yarl==1.7.2
zipp==3.6.0
